\name{ChIPseqSet-class}
\Rdversion{1.1}
\docType{class}
\alias{ChIPseqSet-class}
\alias{chipVals<-}
\alias{chipVals<-,ChIPseqSet,matrix-method}
\alias{chipVals}
\alias{chipVals,ChIPseqSet-method}

\title{Class \code{"ChIPseqSet"}}

\description{
  A class for storing count data obtained from ChIP-seq experiments by
  counting the number of reads lying within regions. The class extends
  \code{\linkS4class{eSet}}.
}

\section{Objects from the Class}{
  Objects can be created by calls of the form \code{new("ChIPseqSet",
  assayData, phenoData, featureData, experimentData, annotation,
  protocolData, ...)}. However, usually objects are generated by calling
  \code{\link{summarizeReads}}.
}

\section{Slots}{
  \describe{
    \item{\code{assayData}:}{Object of class
      \code{"AssayData"}. Contains a matrix with the number of ChIP-seq
      reads.}
    \item{\code{phenoData}:}{Object of class
      \code{"AnnotatedDataFrame"}. Contains experimenter-supplied
      variables describing sample (i.e., columns in \code{assayData})
      phenotypes.}
    \item{\code{featureData}:}{Object of class
      \code{"AnnotatedDataFrame"}. Contains variables describing features
      (i.e., rows in \code{assayData}) unique to this experiment.}
    \item{\code{experimentData}:}{Object of class
      \code{"MIAxE"}. Contains details of experimental methods.}
    \item{\code{annotation}:}{Object of class \code{"character"}. Label
      associated with the annotation package used in the experiment.}
    \item{\code{protocolData}:}{Object of class
      \code{"AnnotatedDataFrame"}. Contains ChIP-seq
      equipment-generated variables describing sample (i.e., columns in
      \code{assayData}) phenotypes.}
    \item{\code{.__classVersion__}:}{Object of class
      \code{"Versions"}. Describes the R and Biobase version numbers
      used to created the instance. Intended for developer use.}
  }
}

\section{Extends}{
Class \code{"\linkS4class{eSet}"}, directly.
Class \code{"\linkS4class{VersionedBiobase}"}, by class "eSet", distance 2.
Class \code{"\linkS4class{Versioned}"}, by class "eSet", distance 3.
}

\section{Methods}{
  \describe{
    \item{chipVals<-}{\code{signature(object = "ChIPseqSet", value = "matrix")}: 
      Set element named \code{chipVals} in the \code{AssayData-class}
      slot.}
    \item{chipVals}{\code{signature(object = "ChIPseqSet")}: Access
      element named \code{chipVals} in the \code{AssayData-class}
      slot.}
  }
}

\author{
  Hans-Ulrich Klein (h.klein@uni-muenster.de)
}

\seealso{
  \code{\link{summarizeReads}}
  \code{linkS4class{eSet}}
}

\examples{
showClass("ChIPseqSet")
}

\keyword{classes}
